APP/requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
plotly==5.20.0
faker==24.0.0
snowflake-connector-python==3.7.1
python-dotenv==1.0.1
altair==5.2.0
pyarrow==15.0.0
